2.10 release
2.10 ----------------------------------------------------------------------------------
- Improving Arduino support
- Adding Esp32 support
- Updated the win32 build to VS2022
- separated the wr_run function into it's two parts: allocating the script
  then initializing it by running "function 0" with:
     // wr_run actually calls these two functions back-to back. If you want
     // to separate the process you can call them on your own.
     // !!!! WARNING: If wr_executeFunctionZero() FAILS you must destroy
     // the context with wr_destroyContext() !!!!!!
     WRContext* wr_allocateNewScript( WRState* w, const unsigned char* block, const int blockSize );
     bool wr_executeFunctionZero( WRState* w, WRContext* context );
- Fixed null switch-body crash (who does that? well.. if you comment them ALL out...)
- Rewrote switch to properly handle local space, major bug that it wasn't before. This
  Represents a difference in the way the switch VM functions but actually _is_ reverse compatible,
  ...almost... the (Switch) case has a (uint16_t) cast which prevents it from functioning, so
  I do need to bump the major version darnit.

+++ discrete_src/vm.cpp
@@ -593,7 +593,7 @@ int wr_callFunction( WRState* w, WRContext* context, WRFunction* function, const
 #endif
 
 	w->err = WR_ERR_None;
-
+	
 	if ( function )
 	{
 		stackTop->p = 0;
@@ -1393,21 +1393,22 @@ NextIterator:
 
 				hashLoc += 4; // yup, point hashLoc to jump vector
 				
-				pc += (uint16_t)READ_16_FROM_PC(hashLoc);
+				pc += READ_16_FROM_PC(hashLoc);
 				CONTINUE;
 			}
 
 			CASE(SwitchLinear):
 			{
 				hashLocInt = (--stackTop)->getHash(); // the "hashes" were all 0<=h<256
-				if ( hashLocInt < *pc++ ) // doesn't mean you switched() on one though ;) check against default top
+				
+				if ( hashLocInt < *pc++ ) // catch selecting > size
 				{
 					hashLoc = pc + (hashLocInt<<1) + 2; // jump to vector
-					pc += READ_16_FROM_PC(hashLoc); // and read it
+					pc += READ_16_FROM_PC( hashLoc ); // and read it
 				}
 				else
 				{
-					pc += READ_16_FROM_PC(pc); // already at default vector
+					pc += READ_16_FROM_PC( pc );
 				}
 				CONTINUE;
 			}
